+++ lib/togls/feature_toggle_registry.rb
@@ -7,16 +7,13 @@ module Togls
   # respective entities. This plays a significant portion in the primary DSL as
   # well.
   class FeatureToggleRegistry
-    def initialize(&block)
+    def initialize(feature_repository, &block)
+      @feature_repository = feature_repository
       @toggle_repository_drivers = [
         Togls::ToggleRepositoryDrivers::InMemoryDriver.new,
         Togls::ToggleRepositoryDrivers::EnvOverrideDriver.new]
-      @feature_repository_drivers =
-        [Togls::FeatureRepositoryDrivers::InMemoryDriver.new]
       @rule_repository_drivers =
         [Togls::RuleRepositoryDrivers::InMemoryDriver.new]
-      @feature_repository = Togls::FeatureRepository.new(
-        @feature_repository_drivers)
       @rule_repository = Togls::RuleRepository.new(@rule_repository_drivers)
       @toggle_repository = Togls::ToggleRepository.new(
         @toggle_repository_drivers, @feature_repository, @rule_repository)